Fix bug where redex.body becomes stale when substitute_mut moves the body.
When a redex's body consists of a single substituted leaf node, this causes the
entire body to become garbage, and the old body's parent (aka: the
abstraction) gets repointed to the new argument subtree that gets made.

This was causing issues when we go to repoint the redex's parent to the
body, since we were repointing to the old (now garbage) body.

This means that the value of `redex.body` becomes stale and therefore
should no longer be used. To fix this, we have substitute_mut return the
correct move-or-not-moved value of the new body.